Change location of parseCSV
